BatallaMyM/src/test/java/classes/magias/ataque/ExpelliarmusTest.java
package classes.magias.ataque;

import classes.personajes.Auror;
import interfaces.Combatiente;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

public class ExpelliarmusTest {

    private Expelliarmus hechizo;
    private Combatiente objetivo;

    @BeforeEach
    public void setUp() {
        hechizo = new Expelliarmus();
        objetivo = new Auror("Auror1", true);
    }

    @Test
    public void testEjecutar() {
        hechizo.ejecutar(objetivo);
        assertEquals(10, ((Auror) objetivo).getPuntosDeVida());
    }

    @Test
    public void testGetCoste() {
        assertEquals(9, hechizo.getCoste());
    }

    @Test
    public void testGetDanio() {
        assertEquals(90, hechizo.getDanio());
    }
}